Most finished only missing the having various sketched values.
Added the beginning menu, correctly removes cell based on difficulty, functioning restart button, and making the value darker once clicked enter. The only thing that is missing is that if the user wants to type in multiple sketched values the current code does not handle that. Right now if the user enters another sketched value it is replaced not added.